+++ sklearn/cross_validation.py
@@ -9,7 +9,7 @@ validation and performance evaluation.
 # License: BSD Style.
 
 from itertools import combinations
-from math import ceil, factorial
+from math import ceil, floor, factorial
 import operator
 
 import numpy as np
@@ -672,6 +672,11 @@ class ShuffleSplit(object):
         Should be between 0.0 and 1.0 and represent the proportion of
         the dataset to include in the test split.
 
+    train_fraction : float or None (default is None)
+        Should be between 0.0 and 1.0 and represent the proportion of
+        the dataset to include in the train split. If None, the value is
+        automatically set to the complement of the test fraction.
+
     indices : boolean, optional (default True)
         Return train/test split as arrays of indices, rather than a boolean
         mask array. Integer indices are required when dealing with sparse
@@ -693,9 +698,18 @@ class ShuffleSplit(object):
     >>> for train_index, test_index in rs:
     ...    print "TRAIN:", train_index, "TEST:", test_index
     ...
-    TRAIN: [2 3 1] TEST: [0]
+    TRAIN: [3 1 0] TEST: [2]
+    TRAIN: [2 1 3] TEST: [0]
     TRAIN: [0 2 1] TEST: [3]
-    TRAIN: [3 0 2] TEST: [1]
+
+    >>> rs = cross_validation.ShuffleSplit(4, n_iterations=3,
+    ...     train_fraction=0.5, test_fraction=.25, random_state=0)
+    >>> for train_index, test_index in rs:
+    ...    print "TRAIN:", train_index, "TEST:", test_index
+    ...
+    TRAIN: [3 1] TEST: [2]
+    TRAIN: [2 1] TEST: [0]
+    TRAIN: [0 2] TEST: [3]
 
     See also
     --------
@@ -703,21 +717,35 @@ class ShuffleSplit(object):
     """
 
     def __init__(self, n, n_iterations=10, test_fraction=0.1,
-                 indices=True, random_state=None):
+                 train_fraction=None, indices=True, random_state=None):
         self.n = n
         self.n_iterations = n_iterations
         self.test_fraction = test_fraction
+        self.train_fraction = train_fraction
         self.random_state = random_state
         self.indices = indices
+        if test_fraction >= 1.0:
+            raise ValueError(
+                "test_fraction=%f should be smaller than 1.0" % test_fraction)
+        if (train_fraction is not None
+            and train_fraction + test_fraction > 1.0):
+            raise ValueError(
+                'The sum of train_fraction=%f and test_fraction=%f '
+                'should be smaller or equal than 1.0' %
+                (train_fraction, test_fraction))
 
     def __iter__(self):
         rng = self.random_state = check_random_state(self.random_state)
         n_test = ceil(self.test_fraction * self.n)
+        if self.train_fraction is None:
+            n_train = self.n - n_test
+        else:
+            n_train = floor(self.train_fraction * self.n)
         for i in range(self.n_iterations):
             # random partition
             permutation = rng.permutation(self.n)
-            ind_train = permutation[:-n_test]
-            ind_test = permutation[-n_test:]
+            ind_test = permutation[:n_test]
+            ind_train = permutation[n_test:n_test + n_train]
 
             if self.indices:
                 yield ind_train, ind_test